chore: cleanup and some refactoring.
