+++ openshift/ftest-openshift-assistant-operational-methods/src/test/java/org/arquillian/cube/openshift/operational/methods/OpenShiftAssistantOperationalMethodsIT.java
@@ -2,6 +2,9 @@ package org.arquillian.cube.openshift.operational.methods;
 
 import io.fabric8.kubernetes.api.model.v3_1.ObjectMeta;
 import io.fabric8.openshift.api.model.v3_1.Project;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
 import org.arquillian.cube.openshift.impl.client.OpenShiftAssistant;
 import org.arquillian.cube.openshift.impl.requirement.RequiresOpenshift;
 import org.arquillian.cube.requirement.ArquillianConditionalRunner;
@@ -10,7 +13,10 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
+import java.io.IOException;
+import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -28,34 +34,46 @@ public class OpenShiftAssistantOperationalMethodsIT {
     }
 
     @Test
+    public void should_apply_template_programmatically() throws IOException {
+
+        openShiftAssistant
+                .usingTemplate(getClass().getClassLoader().getResource("hello-template.yaml"))
+                .parameter("RESPONSE", "Hello from Arquillian Template")
+            .deploy();
+
+        final Optional<URL> route = openShiftAssistant.getRoute();
+        openShiftAssistant.awaitUrl(route.get());
+
+        OkHttpClient okHttpClient = new OkHttpClient();
+        Request request = new Request.Builder().get().url(route.get()).build();
+        Response response = okHttpClient.newCall(request).execute();
+
+        assertThat(response).isNotNull();
+        assertThat(response.code()).isEqualTo(200);
+        assertThat(response.body().string()).isEqualTo("Hello from Arquillian Template\n");
+    }
+
+    @Test
     public void should_list_all_projects() {
-        // given
         String currentProject = openShiftAssistant.getCurrentProject();
 
-        // when
         List<Project> projects = openShiftAssistant.listProjects();
 
-        // then
         assertThat(projects).extracting(Project::getMetadata)
-            .hasSize(2)
             .extracting(ObjectMeta::getName).contains(currentProject);
     }
 
     @Test
     public void should_find_project() {
-        // given
         String currentProject = openShiftAssistant.getCurrentProject();
 
-        // then
         assertThat(openShiftAssistant.findProject(currentProject)).isPresent();
     }
 
     @Test
     public void should_check_if_project_exists() {
-        // given
         String currentProject = openShiftAssistant.getCurrentProject();
 
-        // then
         assertThat(openShiftAssistant.projectExists(currentProject)).isTrue();
     }
 }
