--- openshift/ftest-openshift-assistant-template/src/test/resources/hello-template.yaml
@@ -1,48 +1,0 @@
-apiVersion: v1
-kind: Template
-metadata:
-  name: hello
-parameters:
-- name: RESPONSE
-  value: "Hello from Arquillian"
-  required: false
-objects:
-- apiVersion: v1
-  kind: Route
-  metadata:
-    name: hello-openshift-route
-  spec:
-    port:
-      targetPort: 8080
-    to:
-      kind: Service
-      name: hello-openshift-service
-- apiVersion: v1
-  kind: Service
-  metadata:
-    name: hello-openshift-service
-  spec:
-    ports:
-    - name: hello-openshift-service
-      port: 8080
-      protocol: TCP
-      targetPort: 8080
-    selector:
-      name: hello-openshift-service
-- apiVersion: v1
-  kind: Pod
-  metadata:
-    labels:
-      name: hello-openshift-service
-    name: hello-openshift
-  spec:
-    containers:
-    - image: openshift/hello-openshift
-      imagePullPolicy: IfNotPresent
-      name: hello-openshift
-      ports:
-      - containerPort: 8080
-        protocol: TCP
-      env:
-      - name: RESPONSE
-        value: ${RESPONSE}
